Weather Warnings Data Retrieved
Added functionality to get BOM weather warnings from different WW API,
After the data is received, it is filtered based on (a) whether it has a code relevant to marine conditions and (b) whether the text of the warning references the Lancelin or Perth coasts.  These are returned separately so they can be rendered correctly.  
The functions return a structure that contains 2 arrays.  An array of warnings relevant to the Lancelin Coast and another array that is relevant to the Perth Coast.